+++ src/org/helioviewer/jhv/gui/interfaces/JHVTreeNode.java
@@ -1,0 +1,9 @@
+package org.helioviewer.jhv.gui.interfaces;
+
+import java.awt.Component;
+
+public interface JHVTreeNode {
+
+    Component getComponent();
+
+}
