--- src/org/helioviewer/jhv/layers/fov/FOVTreeNode.java
@@ -1,9 +1,0 @@
-package org.helioviewer.jhv.layers.fov;
-
-import java.awt.Component;
-
-interface FOVTreeNode {
-
-    Component getComponent();
-
-}
